backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# Async & HTTP
aiofiles>=23.0
python-multipart>=0.0.6
httpx>=0.24.0

# LLM Providers
openai>=1.0.0                    # OpenAI, DeepSeek, and Grok providers
google-generativeai>=0.3.0       # Google provider
anthropic>=0.7.0                 # Claude/Anthropic provider (if planning to add)
requests>=2.28.0                 # Ollama provider

# File Processing
PyPDF2>=3.0.0
pandas>=1.5.0
chardet>=5.2.0                   # For CSV encoding detection

# Database
sqlalchemy>=2.0
alembic>=1.12.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0
flake8>=6.0
isort>=5.12.0